requirements.txt
# 视频处理核心依赖
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
decorator==4.4.2  # 修改为兼容 moviepy 的版本
proglog==0.1.10
tqdm==4.66.2

# Web UI 依赖
gradio==4.19.2

# 网页自动化依赖
playwright==1.42.0

# 基础工具依赖
python-dateutil==2.8.2